Environments Page - Environments sub-component breakout pt2 - EnvironmentsPage - directory cleanup and code-owners patch for EnvironmentsPage top-level component. - EnvironmentsPage is now moved into it's own subdirectory inside src/pages - no functional changes
